+++ sqllin-dsl-test/src/jvmTest/kotlin/com/ctrip/sqllin/dsl/test/JvmTest.kt
@@ -79,6 +79,9 @@ class JvmTest {
     @Test
     fun testStringOperators() = commonTest.testStringOperators()
 
+    @Test
+    fun testEnumOperations() = commonTest.testEnumOperations()
+
     @BeforeTest
     fun setUp() {
         deleteDatabase(path, CommonBasicTest.DATABASE_NAME)